% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudr-package.R
\docType{package}
\name{sudr-package}
\alias{sudr}
\alias{sudr-package}
\title{sudr: Data to support the book, Sustainable Urban Development with R}
\description{
The package accompanies the book, Sustainable Urban Development with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.hughbartling.com}
}

}
\author{
\strong{Maintainer}: Hugh Bartling \email{hugh.bartling@depaul.edu}

}
\keyword{internal}
